Match wcfloortile_update u8 alpha store conversion
The float->u8 alpha store (objAnim->alpha = t) needs an explicit
(u8)(int) cast to reproduce the target's clrlwi r0,r0,24 mask before
the stb. wcfloortile_update 98.357 -> 98.826.

+++ src/main/dll/WC/dll_0298_wcfloortile.c
@@ -162,7 +162,7 @@ void wcfloortile_update(int obj)
                 }
                 t = t * lbl_803E6EB0;
             }
-            objAnim->alpha = t;
+            objAnim->alpha = (u8)(int)t;
         }
         if (objAnim->alpha == 0)
         {
